ecs:service-create tweaks
